requirements.txt
numpy
pandas
pytorch_lightning==2.4.0
torch
torchmetrics